+++ docs/appendix/version_history.rst
@@ -1,6 +1,15 @@
 Version History
 ===============
 
+Version 2.2
+-----------
+
+Version 2.2.0 GA (2019-04-21)
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+* Add :c:func:`ms3_init` to replace :c:func:`ms3_thread_init` and deprecate the latter.
+* Add :c:func:`ms3_library_init_malloc` to add custom allocators
+
 Version 2.1
 -----------